genai_agent_project/requirements.txt
redis>=4.0.0,<5.0.0
pyyaml>=5.1,<7.0
aiohttp>=3.7.0,<4.0.0
